Reorganizing fragment inheritance hierarchy and messaging
Summary:
Majorly refactored the upcoming events tab fragment into three fragments, one an abstract class that is the general superclass
of all lists of events, one that inherits from that specifically for the upcoming event list, and one that contains the list
in order to also have the FAB
This will make making a tab for past attended events significantly easier, and is probably a much better design pattern overall.
Hooked through the messaging to catch touch events and trigger an appropriate dialog box, which i used the edit fragment for
(temporarily)
Also fixed the lingering bug of new event failures because it turns out the local datastore does NOT give a parse object
a object id.
Therefore the role of creating a new event object is now the responsibility of the edit event fragment which is how it ought to
have been originally

Test Plan: build and run, adding events through the add menu, and then pulling up the whole view by clicking on the event

Reviewers: connie, mariazlatkova, kaogra

Reviewed By: mariazlatkova, kaogra

Differential Revision: https://phabricator.fb.com/D2233207

Tasks: 7667347, 7663306, 7667344

Signature: t1:2233207:1436540270:dbbaa2b7d6277f350f156b3b1e3d6aa99653f7ea